Rename module tutorial to _tutorial
Also update other modules
