+++ tests/test_index_search.py
@@ -6,7 +6,7 @@ import xgcm
 from parcels._datasets.structured.generic import datasets
 from parcels._index_search import _search_indices_curvilinear_2d
 from parcels.field import Field
-from parcels.tutorial import download_example_dataset
+from parcels._tutorial import download_example_dataset
 from parcels.xgrid import XGrid
